pyproject.toml
[project]
name = "api-testing-agent"
version = "1.0.0"
description = "Automated API test generation and execution service using LLM"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "requests",
    "pygithub",
    "openai",
    "python-dotenv",
    "pydantic",
    "pyyaml",
    "pytest",
    "schemathesis",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39']

[project.scripts]
api-testing-agent = "main:main"
